Add allocateCachePage support
Initializes all buffer and cache blocks to be set to available and volatile or
not depending on the storage type.

When allocating a new cache page, bufman first looks for any available volatile
block slots and places the virtual diskaddress there. If there are no available
slots, a volatile slot is randomly picked and the data in that slot is evicted
to the persistent buffer. The virtual diskaddress is then placed in the evicted
cache slot.