Compare pageExistsInBuffer with SUCCESS
The return value of pageExistsInBuffer was previously being used on it's own
if conditions. This ended up causing a problem due to SUCCESS being equal to 0
and ERROR being equal to -1.

This meant that if (pageExistsInBuffer) would be interpreted as false when it
returned SUCCESS. This changes those statements to compare directly to SUCCESS.