Add createPersistentTable and createVolatileTable functions
The createPersistentTable function requires some of the readwrite layer
functionality, so I've left that part out for now. I am also not sure about
where to store some of the table information, so I left some TODOs with some of
the questions I wasn't able to come up with an answer for.