Throw in initial implementations of read,write,put/getPage
I wrote read and write as special cases of get/putPage because I thought it
would be nice to encapsulate buffer access through only getPage and putPage.

Someone during lecture also mentioned using a `char *` as an out parameter during
class instead of returning a char * of the data read. I thought that was a
little easier on read, write, getPage, and putPage because those function don't
have to worry about allocating memory for the bytearray that they return and
can let the caller handle it.