Add generateRecordDescription function
This function takes a tableDescription and two out parameters for the actual
buffer and the record size. It writes the column name into the buffer, followed
by a type byte and a size byte if the type is a varchar.

The DBMS can use this buffer to understand the order, names, types, and sizes
of each attribute in the record.

This still needs some work to make sure padding is added to the record
description and record size correctly.